equals signs, path name
