+++ Screenshot-App/index.html
@@ -4,21 +4,20 @@
 <head>
 	<meta charset="UTF-8">
 	<title>Screenshot</title>
+	<meta name="viewport" content="width=device-width, initial-scale=1.0">
 	<link rel='stylesheet' href='./css/photon.min.css'>
-	<script>if (typeof module === 'object'){window.module = module; module = undefined;}</script>
-	<script type="text/javascript" src = "renderer.js"></script>
-	<script>if (window.module) module = window.module;</script>
+
 </head>
 
 <body>
+
 	<div class="window">
 		<label id='screenshot-path'>Path:</label>
 		<button id='screen-shot' class = "btn btn-primary">Take Screenshot</button>
 	</div>
-
+	<script>if (typeof module === 'object'){window.module = module; module = undefined;}</script>
+	<script type="text/javascript" src="renderer1.js"></script>
+	<script>if (window.module) module = window.module;</script>
 
 </body>
-
-
-	
 </html>
